Moved formatMessage to ndla-ui/i18n. Removed ndla/i18n package
