--- packages/ndla-i18n/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@ndla/i18n",
-  "version": "0.5.6",
-  "description": "NDLA's own i18n package based on react-intl",
-  "license": "GPL-3.0",
-  "main": "lib/index.js",
-  "module": "es/index.js",
-  "types": "lib/index.d.ts",
-  "sideEffects": false,
-  "scripts": {
-    "build": "yarn build:types",
-    "build:types": "tsc -p tsconfig.build.json",
-    "prepublish": "yarn build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/NDLANO/frontend-packages.git",
-    "directory": "packages/ndla-i18n"
-  },
-  "keywords": [
-    "ndla"
-  ],
-  "author": "ndla@knowit.no",
-  "files": [
-    "lib",
-    "es"
-  ],
-  "peerDependencies": {
-    "prop-types": "^15.6.0",
-    "react": "^15.6.0 || ^16.8.0"
-  },
-  "dependencies": {
-    "@types/hoist-non-react-statics": "^3.3.1",
-    "hoist-non-react-statics": "^3.3.0",
-    "i18next": "^20.3.1",
-    "i18next-browser-languagedetector": "^6.1.1",
-    "intl-format-cache": "^4.3.1",
-    "intl-messageformat": "5.4.3",
-    "invariant": "^2.2.3",
-    "react-i18next": "^11.11.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
