--- packages/ndla-i18n/src/t.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) 2016-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-export default interface tType {
-  t(arg: string, obj?: { [key: string]: string | boolean | number }): string;
-}
