--- packages/ndla-i18n/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "baseUrl": "./",
-    "declarationDir": "./lib",
-    "rootDir": "./src"
-  },
-  "include": ["./src"]
-}
