+++ packages/ndla-ui/package.json
@@ -45,13 +45,18 @@
     "@ndla/util": "^0.4.8",
     "@reach/menu-button": "^0.12.1",
     "@reach/slider": "^0.12.1",
+    "@types/hoist-non-react-statics": "^3.3.1",
     "@types/react-syntax-highlighter": "^11.0.4",
     "classnames": "2.2.6",
     "focus-trap": "^4.0.2",
     "focus-trap-react": "^6.0.0",
+    "hoist-non-react-statics": "^3.3.0",
     "html-react-parser": "^0.10.3",
     "i18next": "^20.3.1",
     "i18next-browser-languagedetector": "^6.1.1",
+    "intl-format-cache": "^4.3.1",
+    "intl-messageformat": "5.4.3",
+    "invariant": "^2.2.3",
     "jump.js": "^1.0.2",
     "lodash": "^4.17.20",
     "react-bem-helper": "1.4.1",
