+++ packages/ndla-ui/src/i18n/formatMessage.ts
